MOVEONLY: Libconsensus: Move consensus functions away from main.cpp
