+++ src/Makefile.am
@@ -179,6 +179,7 @@ libbitcoin_server_a_SOURCES = \
   bloom.cpp \
   chain.cpp \
   checkpoints.cpp \
+  consensus/consensus.cpp \
   httprpc.cpp \
   httpserver.cpp \
   init.cpp \
