+++ src/wallet/walletdb.cpp
@@ -6,8 +6,8 @@
 #include "wallet/walletdb.h"
 
 #include "base58.h"
+#include "consensus/consensus.h"
 #include "consensus/validation.h"
-#include "main.h" // For CheckTransaction
 #include "protocol.h"
 #include "serialize.h"
 #include "sync.h"
